pyproject.toml
[project]
name = "indian-ai-hedge-fund"
version = "0.1.0"
description = ""
authors = [
    {name = "Amit Wani",email = "mtwn105@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "kiteconnect (>=5.0.1,<6.0.0)",
    "agno (>=1.3.3,<2.0.0)",
    "rich (>=14.0.0,<15.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "yfinance (>=0.2.55,<0.3.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "openai (>=1.75.0,<2.0.0)",
    "langchain (>=0.3.23,<0.4.0)",
    "langchain-openai (>=0.3.14,<0.4.0)",
    "google-genai (>=1.11.0,<2.0.0)",
    "colorama (>=0.4.6,<0.5.0)",
]

[tool.poetry]
packages = [{include = "indian_ai_hedge_fund", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
